pyproject.toml
[project]
name = "tinker-ki-starter"
version = "0.1.0"
description = ""
authors = [
    {name = "Jonas Arndt",email = "joebubi@posteo.de"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "signalbot >=1.0.1,<2.0.0",
    "python-dotenv >=1.2.2,<2.0.0",
    "faster-whisper >=1.2.1,<2.0.0",
    "ollama >=0.6.1,<0.7.0",
    "ddgs >=9.14.1,<10.0.0",
    "pathlib >=1.0.1,<2.0.0",
    "opentelemetry-api>=1.41.1",
    "opentelemetry-sdk>=1.41.1",
    "opentelemetry-exporter-otlp>=1.41.1",
    "loguru>=0.7.3",
]
